Fixed a whole bunch of printf/scanf format/argument combinations. - %o and %x require an argument of type unsigned; - %n requires a pointer to (signed) int (%zn to ssize_t); - an unsigned short (including SQLUSMALLINT), when used in an expression   (including as argument) is automatically converted to (signed) int. Some other small fixes as well.

+++ clients/mapiclient/mclient.c
@@ -691,10 +691,10 @@ XMLprdata(const char *val)
 			default:
 				if ((codepoint & ~0x80) <= 0x1F || codepoint == 0177) {
 					/* control character */
-					mnstr_printf(toConsole, "&#%d;", codepoint);
+					mnstr_printf(toConsole, "&#%u;", codepoint);
 				} else if (codepoint < 0x80) {
 					/* ASCII */
-					mnstr_printf(toConsole, "%c", codepoint);
+					mnstr_printf(toConsole, "%c", (int) codepoint);
 				} else {
 					mnstr_printf(toConsole, "&#x%x;", codepoint);
 				}
